pyproject.toml
[tool.poetry]
name = "nclone"
version = "0.1.0"
description = "An N++ simulator with headless mode support"
authors = ["SimonV42 <simonv42@example.com>"] # Placeholder email
readme = "README.md"
packages = [{include = "nclone", from = "src"}]
homepage = "https://github.com/SimonV42/nclone"

[tool.poetry.dependencies]
python = ">=3.9"
pycairo = "*"
numpy = "*"
imageio = {version = "*", extras = ["ffmpeg"]}
pyglet = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
